refactor(daemon): give the memory tools a domain home, invert memory -> mcp
`src/memory/` no longer imports the MCP transport module. The memory tool
descriptors (`MEMORY_TOOLS`, `MEMORY_TOOL_NAMES`, the external-memory record
tools and `externalMemoryTools()`) move to `src/memory/tools.ts`, and
`MEMORY_ACCESS_BLOCKED` moves there too — the refusal is memory-access policy,
not tool dispatch. `src/mcp/tools.ts` and `src/mcp/ops.ts` now import them
(transport may know the domain).

`ToolDescriptor` and its schema shapes are the MCP `tools/list` wire shape used
by mcp/, platforms/ and evaluation/ as well as memory/, so they move to a new
leaf `src/mcp/tool-descriptor.ts` with no internal imports rather than being
hosted in the memory domain. Every importer is retargeted; no re-export shims
are left behind.
